Fix pluginManager when app restarts
